chore(query): tidy aggregate refactor fallout
